+++ TidalPy/utilities/math/special_x.pyx
@@ -1,4 +1,4 @@
-from libc.math cimport tgamma
+from libc.math cimport tgamma, NAN
 
 cdef double[51] pre_calculated_doubles
 cdef double* pre_calculated_doubles_ptr = &pre_calculated_doubles[0]
@@ -55,7 +55,7 @@ pre_calculated_doubles_ptr[ 49] = 58435841445947288807899666579456.00
 pre_calculated_doubles_ptr[ 50] = 520469842636666553028024352112640.00
 
 
-cdef double cf_double_factorial(unsigned char n) nogil:
+cdef double cf_double_factorial(unsigned char n) noexcept nogil:
 
     if n < 51:
         # Use precalculated doubles
@@ -64,8 +64,13 @@ cdef double cf_double_factorial(unsigned char n) nogil:
         # Calculate using recursion
         return tgamma(<double>n + 1.) / cf_double_factorial(n - 1)
     else:
-        raise ValueError('C function `tgamma` experiences overflow for l > 170.')
+        # ValueError('C function `tgamma` experiences overflow for l > 170.')
+        return NAN
 
 
 def double_factorial(unsigned char n):
+    
+    if n >= 171:
+        raise ValueError('C function `tgamma` experiences overflow for l > 170.')
+
     return cf_double_factorial(n)